Move EditDocHistoryController to ExtendedController
