+++ Core/Lib/ExtendedController/EditDocHistoryController.php
@@ -16,9 +16,7 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-namespace FacturaScripts\Core\Controller\Base;
-
-use FacturaScripts\Core\Lib\ExtendedController\EditController;
+namespace FacturaScripts\Core\Lib\ExtendedController;
 
 /**
  * Controller for editing models that are related and show
